refactor: :zap: Updated types
Updated types for cards
