+++ components/molecules/cards/types.ts
@@ -1,0 +1,9 @@
+import { Film, Person, Specie } from "@/lib/types";
+
+export interface CardProps {
+  isFav: boolean;
+  handleUpdateFavourite: (
+    item: Film | Person | Specie,
+    isFavourite: boolean
+  ) => Promise<void>;
+}
