+++ components/organisms/cardList/cardList.tsx
@@ -1,17 +1,9 @@
-import {
-  CardLoader,
-  FilmCard,
-  FilmCardProps,
-  Header,
-  PersonCard,
-  PersonCardProps,
-  SpecieCard,
-  SpecieCardProps,
-} from "@/components/molecules";
+import { CardLoader, Header } from "@/components/molecules";
 import { GetFavourites } from "@/hooks/useFavourite";
 import { Data, Film, Person, Specie } from "@/lib/types";
 import { SerializedError } from "@reduxjs/toolkit";
 import { FetchBaseQueryError } from "@reduxjs/toolkit/dist/query";
+import { List } from "./list";
 
 type ListItemProps = {
   title: string;
@@ -89,61 +81,5 @@ export const CardList = ({
         ) : null}
       </div>
     </div>
-  );
-};
-
-type ListProps = {
-  list: (Film | Person | Specie)[];
-  favourites: GetFavourites;
-  handleUpdateFavourite: (
-    item: Film | Person | Specie,
-    isFavourite: boolean
-  ) => Promise<void>;
-};
-
-interface Views {
-  film: ({ item, isFav, handleUpdateFavourite }: FilmCardProps) => JSX.Element;
-  person: ({
-    item,
-    isFav,
-    handleUpdateFavourite,
-  }: PersonCardProps) => JSX.Element;
-  specie: ({ item, isFav }: SpecieCardProps) => JSX.Element;
-}
-
-const views: Views = {
-  film: FilmCard,
-  person: PersonCard,
-  specie: SpecieCard,
-};
-
-const List = ({ list, favourites, handleUpdateFavourite }: ListProps) => {
-  const { isLoading, isFetching, data } = favourites;
-  return (
-    <>
-      {list.map((item: Film | Person | Specie) => {
-        let exists;
-
-        const CurrentView = views[item._type as keyof Views];
-
-        if (!isLoading && !isFetching && data) {
-          exists = data.find(
-            (el: Film | Person | Specie) => item.url === el.url
-          );
-        }
-
-        return (
-          <div key={item.url}>
-            {CurrentView && (
-              <CurrentView
-                item={item as any}
-                isFav={exists !== undefined}
-                handleUpdateFavourite={handleUpdateFavourite}
-              />
-            )}
-          </div>
-        );
-      })}
-    </>
   );
 };
